feat: Enhance robustness with BigDecimal for monetary calculations, resilient fallback strategy using cache, async notifications

+++ src/main/java/com/mycompany/currency/converter/Presentation/DTO/ConversionResponse.java
@@ -3,15 +3,16 @@ package com.mycompany.currency.converter.Presentation.DTO;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
 @Builder
 public class ConversionResponse {
     private LocalDate conversionDate;
-    private Double exchangeRate;
-    private Double originalAmount;
-    private Double convertedAmount;
+    private BigDecimal exchangeRate;
+    private BigDecimal originalAmount;
+    private BigDecimal convertedAmount;
     private String sourceCurrency;
     private String targetCurrency;
 }
